requirements.txt
aioredis==1.2.0
asgi-redis==1.0.0
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
channels==2.2.0
channels-redis==2.4.0
constantly==15.1.0
daphne==2.3.0
Django==2.2.1
django-crispy-forms==1.7.2
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.1
msgpack-python==0.5.6
Pillow==6.0.0
PyHamcrest==1.9.0
pypiwin32==223
pytz==2018.5
pywin32==224
redis==2.10.6
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.0
txaio==18.8.1
virtualenv==16.4.3
zope.interface==4.6.0
